requirements.txt
#python>=3.9
numpy>=1.23.1
scipy
astropy==5.3.3
aplpy
matplotlib<3.6
scikit-image
astroquery
xmltodict
python-casacore
pandas
